add sat score data, remove unused classes and files, separate out composables
