+++ app/build.gradle
@@ -17,6 +17,9 @@ android {
         versionName "1.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+
+        buildConfigField("String", "appToken", appToken)
+
     }
 
     buildTypes {
@@ -67,6 +70,7 @@ dependencies {
     implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.6.1'
     implementation 'androidx.fragment:fragment-ktx:1.5.7'
     implementation 'androidx.activity:activity-compose:1.7.2'
+    implementation "androidx.lifecycle:lifecycle-viewmodel-compose:2.6.1"
 
     implementation platform('androidx.compose:compose-bom:2023.05.01')
     implementation 'androidx.compose.ui:ui'
@@ -76,6 +80,7 @@ dependencies {
     implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.7.1'
     implementation 'org.jetbrains.kotlin:kotlin-stdlib-jdk7:1.8.21'
     implementation 'androidx.test.ext:junit-ktx:1.1.5'
+    implementation 'androidx.compose.runtime:runtime-livedata:1.5.0-beta01'
 
     kapt 'com.google.dagger:hilt-compiler:2.46.1'
